# web/openapi.yaml
openapi: 3.0.3
info:
  title: Pomodorough API
  version: 1.5.0
  description: |
    Local-first Pomodoro timer synchronization API.

    Timer, task, duration preference, and auto-start preference mutations are immutable and idempotent.
    The server orders them by hybrid logical clock, device ID, and operation ID,
    then returns canonical timer, task, duration, and auto-start projections. Clients must
    retain unacknowledged mutations and retry them with the same IDs.

    Web clients authenticate with the session cookie created by Google OAuth.
    Cookie-authenticated mutations must also send the exact public origin and an
    `X-CSRF-Token` matching the token returned by `GET /api/v1/me`. Native and
    desktop clients authenticate with an access-token bearer credential and do
    not send CSRF credentials.

    Authenticated `/me` user profiles and bootstrap/sync snapshots include
    `accountIncarnation`, an opaque account-lifetime identifier. Public user IDs
    can repeat after deletion; incarnations cannot. Web clients bind cached state
    to this identifier and send `X-Pomodorough-Account-Incarnation` on subsequent
    authenticated requests. A present header must exactly match the authenticated
    account incarnation; otherwise the server returns 409 without executing the
    request. Omission remains supported for existing native clients. Web clients
    must revalidate identity and use Core bootstrap policy on incarnation changes,
...

...
servers:
...
paths:
  /openapi.yaml:
    get:
      tags: [System]
      operationId: getOpenAPISpec
      summary: Download OpenAPI specification
      security: []
      responses:
        "200":
          description: OpenAPI 3.0 specification for this service.
          content:
            application/yaml:
              schema:
                type: string
        "503":
          description: Specification file is unavailable.
          content:
            text/plain:
              schema:
                type: string
...
  /auth/google/start:
    get:
      tags: [Web authentication]
      operationId: startGoogleAuthentication
      summary: Start web Google OAuth
      description: |
        Creates short-lived OAuth state, nonce, and PKCE verifier data in an
        encrypted secure cookie, then redirects to Google. Unsafe or external
        return paths are replaced with `/`.
      security: []
      parameters:
        - name: return
          in: query
          required: false
          description: Same-origin path to open after successful authentication.
          schema:
            type: string
            maxLength: 1024
            default: /
            example: /
      responses:
        "302":
          description: Redirect to Google authorization endpoint.
          headers:
            Location:
              description: Google authorization URL.
              schema:
                type: string
                format: uri
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/PlainInternalServerError"
        "503":
          $ref: "#/components/responses/PlainAuthUnavailable"
  /auth/google/callback:
    get:
      tags: [Web authentication]
      operationId: completeGoogleAuthentication
      summary: Complete web Google OAuth
      description: |
        Google OAuth callback. Validates encrypted state, PKCE, nonce, issuer,
        audience, and verified email; creates a 30-day web session; then sets
        secure session and CSRF cookies.
      security: []
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "303":
          description: Authentication succeeded; redirect to requested local path.
          headers:
            Location:
              description: Validated same-origin return path.
              schema:
                type: string
        "400":
          $ref: "#/components/responses/PlainInvalidAuthState"
        "401":
          $ref: "#/components/responses/PlainAuthenticationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/PlainInternalServerError"
        "502":
          $ref: "#/components/responses/PlainAuthenticationFailed"
        "503":
          $ref: "#/components/responses/PlainAuthUnavailable"
  /api/v1/auth/google/challenge:
    post:
      tags: [Native authentication]
      operationId: createGoogleChallenge
      summary: Create native Google Sign-In challenge
      description: |
        Returns an encrypted five-minute challenge and nonce. Native clients must
        include this nonce when requesting a Google ID token, then submit both
        values to the exchange endpoint.
      security: []
      responses:
        "200":
          description: Challenge created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NativeChallengeResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/auth/google/exchange:
    post:
      tags: [Native authentication]
      operationId: exchangeGoogleIDToken
      summary: Exchange native Google identity for app tokens
      description: |
        Verifies challenge, Google ID-token nonce, issuer, audience, authorized
        party, and verified email. Returns a 15-minute access token and rotating
        30-day refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NativeExchangeRequest"
      responses:
        "200":
          description: Native session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/AuthUnavailable"
  /api/v1/auth/refresh:
    post:
      tags: [Native authentication]
      operationId: refreshNativeTokens
      summary: Rotate native access and refresh tokens
      description: |
        Consumes the supplied refresh token and returns a new token pair. Reusing
        an already consumed refresh token revokes its entire session family.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          description: Refresh token is invalid, expired, revoked, or reused.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid refresh token
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/me:
    get:
      tags: [Account]
      operationId: getCurrentUser
      summary: Get authenticated user and CSRF token
      description: |
        For web sessions, returns the CSRF token required by mutation endpoints
        and replaces it when the CSRF cookie is missing or invalid. `csrfToken`
        is an empty string for bearer-authenticated clients.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Current user session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/logout:
    post:
      tags: [Account]
      operationId: logout
      summary: Revoke current session
      description: |
        Revokes the current web or native session and all tokens belonging to it.
        Web clients must supply both conditional CSRF headers.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204":
          description: Session revoked. Web session cookies are cleared.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/account:
    delete:
      tags: [Account]
      operationId: deleteAccount
      summary: Permanently delete the authenticated account
      description: |
        Permanently removes the account database, including profile, sessions,
        tokens, timers, history, tasks, and preferences. The
        literal confirmation `DELETE` is required. Web clients must also supply
        the conditional CSRF headers. Deletion is immediate in live storage;
        encrypted infrastructure backups may retain deleted data until their
        documented rotation period expires.
        If a successful response is lost, repeat this DELETE with the exact
        originally authorized access token or web session and confirmation.
        A private durable receipt confirms only that deleted account generation,
        even after credential expiry or server restart; it never deletes a
        recreated account or authenticates another endpoint. Web retries require
        the original CSRF token, matching cookie/header, and same-site Origin.
        A 204 is sent only after the old generation's remaining live files are
        purged. A 401 without a receipt is not evidence of successful deletion.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "204":
          description: The credential's account generation was deleted, or its committed deletion was confirmed again.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/revoke-device:
    post:
      tags: [Account]
      operationId: revokeDevice
      summary: Revoke all sessions for a device
      description: |
        Revokes the device and all native sessions and tokens associated with it.
        Web clients must supply both conditional CSRF headers.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeDeviceRequest"
      responses:
        "204":
          description: Matching device sessions revoked. Unknown IDs are a no-op.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/bootstrap:
    get:
      tags: [Timer]
      operationId: getBootstrapSnapshot
      summary: Materialize canonical server state for bootstrap
      description: |
        Returns the same complete canonical snapshot shape as synchronization,
        with empty acknowledgement arrays. It does not create or update a device
        or store a bootstrap resolution. Before returning, it transactionally
        reduces immutable timer commands at current server time and persists a
        changed deadline projection. Such materialization increments revision once
        and publishes that revision; an unchanged projection performs no mutation.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Current canonical account snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Deadline materialization requires a revision beyond maximum safe integer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/bootstrap/resolve:
    post:
      tags: [Timer]
      operationId: resolveBootstrapHistory
      summary: Resolve local and remote bootstrap history
      description: |
        Resolves bootstrap history under an account-scoped lock and transaction.
        `expectedRevision` is compared before first execution. A stale revision
        returns `409 revision conflict` without mutation.

        `keep_remote` requires all operation arrays to be empty and records the
        returned remote snapshot. It ordinarily preserves revision, but may
        transactionally materialize a changed deadline projection and increment
        revision once in the same resolution. `merge` applies the
        supplied immutable local operations on top of remote history.
        `replace_remote` atomically clears timer commands and projections, task
        operations, and duration operations before applying local history. When
        `autoStartOperations` is omitted, replacement preserves remote auto-start
        history for compatibility with legacy clients. A present empty array
        explicitly replaces auto-start history with default `false`.
        `selectedTaskOperations` has the same compatibility behavior: omission
        preserves remote selected-task history, while a present empty array
        explicitly replaces it with no selected task.

        `requestId` provides durable idempotency. An exact retry returns the exact
        stored canonical result without another revision increment or event. Reuse
        with a changed normalized payload returns `409 request ID conflict`.
        A changed merge or replacement increments revision exactly once and
        publishes one revision event only on first execution. Replacing populated
        remote history with empty local history counts as a change.

        Bearer-authenticated clients may only resolve the device bound to their
        access-token session. Web clients must supply both conditional CSRF
        headers.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapResolutionRequest"
      responses:
        "200":
          description: Resolution recorded and canonical snapshot returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "400":
          description: Envelope, strategy, or any supplied operation is invalid; no mutation is stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid bootstrap resolution request
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: CSRF validation failed or bearer session device mismatched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Expected revision is stale, request ID was reused with a changed normalized payload, or canonical revision is exhausted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                revisionConflict:
                  value:
                    error: revision conflict
                requestIDConflict:
                  value:
                    error: request ID conflict
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/sync:
    post:
      tags: [Timer]
      operationId: syncTimer
      summary: Submit commands and read canonical timer state
      description: |
        Accepts up to 256 immutable timer commands, 256 immutable task operations,
        256 immutable duration operations, 256 immutable auto-start operations,
        and 256 immutable selected-task operations.
        Repeating a stored ID is idempotent
        only when immutable payload and device ID match. A changed stored payload
        is rejected. Duplicate timer command, task operation, duration operation,
        auto-start operation, or selected-task operation IDs within one request
        invalidate the entire request before mutation.
        Reusing a timer device sequence for a different command produces a
        rejected acknowledgement. Mutations are ordered by `hlcWallMs`,
        `hlcCounter`, device ID, then operation ID; latest valid intent wins.
        Duration preferences resolve independently for each phase. Auto-start
        operations resolve as one account preference with default `false`.
        Selected-task operations resolve as one nullable account preference. Its
        winning non-null task ID projects as `null` while that task is absent;
        immutable preference history remains unchanged so selection reappears if
        the task later becomes live again.

        `lastRevision` reports the client's last observed revision. This version
        always returns the complete canonical snapshot and history. Revision is a
        monotonic canonical-generation number: it increments once when a request
        inserts one or more operations or when time materializes a changed timer
        projection, such as deadline completion. Equal revision therefore denotes
        the same canonical snapshot generation and needs no wall-clock tie-break.

        Bearer-authenticated clients may only sync the device bound to their
        access-token session. Web clients must supply both conditional CSRF
        headers.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Mutations reconciled and canonical snapshot returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "400":
          description: Envelope or any mutation is invalid; no mutation is stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid sync request
        "409":
          description: Canonical revision reached maximum safe integer and cannot advance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: revision exhausted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: CSRF validation failed or bearer session device mismatched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                csrf:
                  value:
                    error: forbidden
                deviceMismatch:
                  value:
                    error: device mismatch
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/history:
    get:
      tags: [Timer]
      operationId: getTimerHistory
      summary: Get full timer history
      description: |
        Returns completed, cancelled, and superseded timer sessions, newest first.
        History is rebuilt from the immutable command audit log.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Full timer history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Deadline materialization requires a revision beyond maximum safe integer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-pomodorough_session
      description: Secure HttpOnly web session cookie created by Google OAuth.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Native 15-minute access token returned by exchange or refresh.
...
  responses:
    RateLimited:
      description: Request rate or concurrent stream limit exceeded.
      headers:
        Retry-After:
          description: Seconds until the client should retry.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: rate limit exceeded
...
    PlainAuthUnavailable:
      description: Required Google OAuth client configuration is missing.
      content:
        text/plain:
          schema:
            type: string
          example: Google authentication unavailable
    PlainInternalServerError:
      description: Internal server error.
      content:
        text/plain:
          schema:
            type: string
          example: Internal Server Error
    PlainInvalidAuthState:
      description: OAuth state cookie, state query, or authorization code is invalid.
      content:
        text/plain:
          schema:
            type: string
          example: Invalid authentication state
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
...